requirements.txt
# Google Gmail API (for reading email alerts)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Playwright scraping (fallback)
playwright>=1.40.0
playwright-stealth>=1.0.6

# HTTP
requests>=2.31.0

# Config & env
PyYAML>=6.0
python-dotenv>=1.0.0
